current vn is sub vn

+++ Camera.h
@@ -3,8 +3,7 @@ class Camera
 public:
     Camera();
     virtual ~Camera();
-    
-    static ray shootRay(int i, int j);
+    static ray shootRay(float i, float j);
     static void initCamera(float * values);
     static void init(float * values);
 };
